Add experimental remote MDX support to esbuild loader
This adds support for loading MD(X) files over the wire (`https?://`).
This might be interesting for folks who have files in other repos for example.
Or that want to include some readme from GitHub and such.

Note though that this is a) incredibly unsafe: it evals remote code, and b)
very experimental!
Please test it out, see whether this is a good idea or not, and leave feedback!