Update `loader-utils`
This, uhm, updates webpack’s `loader-utils` to version 3, which removes the one
option we were using it for, so this also removed `loader-utils`.
We were using `getOptions`, which is now provided by webpack since v5.
I *believe* webpack 4 never worked with xdm so it shouldn’t affect anyone.
Please open an issue if someone is affected.